I added the add category activity, with some code for converting textView into a string and made a list to keep track of categories with the amount values.
This time I connected the getCategory function to the onClick event of the Save button. The screen will reset after saving.

+++ app/src/main/java/edu/byui/budgetman/model/Transaction.java
@@ -9,10 +9,10 @@ public class Transaction {
     // Only getter method will be provided for this since it will always default to current date
     // from the constructor
     private Date date; // rolled backwards to Date instead of LocalDate for API compatibility
-    private BigDecimal amount;
+    private Float amount;
     private String description;
 
-    public Transaction(BigDecimal amount, String description) {
+    public Transaction(Float amount, String description) {
 
         // When a Transaction is created it defaults its date to the current date
         this.date = new Date();
@@ -25,11 +25,11 @@ public class Transaction {
         return date;
     }
 
-    public BigDecimal getAmount() {
+    public Float getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Float amount) {
         this.amount = amount;
     }
